fix type

+++ src/Button/Button.c
@@ -1,14 +1,14 @@
 #include "Button.h"
 #include "../Test/Test.h"
 
-Button* allocButtonText(SDL_Window* window, SDL_Renderer* renderer, const uint_fast16_t w, const int_fast16_t h, const char * text, TTF_Font * font, const SDL_Color colorText, const SDL_Color back) {
+Button* allocButtonText(SDL_Window* window, SDL_Renderer* renderer, const int w, const int h, const char * text, TTF_Font * font, const SDL_Color colorText, const SDL_Color back) {
     Button* bouton = malloc(sizeof(Button));
     Test( window, renderer, ERROR, bouton == NULL, "malloc", __FILE__, __LINE__);
     loadButtonText( window, renderer, w, h, text, font, bouton, colorText, back);
     return bouton;
 }
 
-void loadButtonText(SDL_Window* window, SDL_Renderer* renderer, const uint_fast16_t w, const uint_fast16_t h, const char * text, TTF_Font * font, Button* bouton, const SDL_Color colorText, const SDL_Color back) {
+void loadButtonText(SDL_Window* window, SDL_Renderer* renderer, const int w, const int h, const char * text, TTF_Font * font, Button* bouton, const SDL_Color colorText, const SDL_Color back) {
     loadText( window, renderer, text, font, &bouton->textButton, colorText, SHADED, &back);
     if (bouton->textButton.RectText->w < (int)w) {
         bouton->rectButton.w = w;
@@ -23,7 +23,7 @@ void loadButtonText(SDL_Window* window, SDL_Renderer* renderer, const uint_fast1
     bouton->colorButton = back;
 }
 
-void printButton(SDL_Window* window, SDL_Renderer* renderer, Button* bouton, const uint_fast16_t x, const uint_fast16_t y) {
+void printButton(SDL_Window* window, SDL_Renderer* renderer, Button* bouton, const int x, const int y) {
     bouton->rectButton.x = x;
     bouton->rectButton.y = y;
     Test( window, renderer, WARNING, SDL_SetRenderDrawColor(renderer, bouton->colorButton.r, bouton->colorButton.g, bouton->colorButton.b, bouton->colorButton.a) != 0, "SDL_SetRenderDrawColor", __FILE__, __LINE__);
@@ -35,7 +35,7 @@ void printButton(SDL_Window* window, SDL_Renderer* renderer, Button* bouton, con
     }
 }
 
-bool clickButton( const Button* bouton, const uint_fast16_t clickX, const uint_fast16_t clickY) {
+bool clickButton( const Button* bouton, const int clickX, const int clickY) {
     if (bouton != NULL) {
         if ((int)clickX >= bouton->rectButton.x && (int)clickX <= bouton->rectButton.x + bouton->rectButton.w && (int)clickY >= bouton->rectButton.y && (int)clickY <= bouton->rectButton.y + bouton->rectButton.h) {
         return true;
