+++ src/Graphic/Graphic.h
@@ -7,7 +7,7 @@
 
 void update_in_game( Game* game);
 
-void pop_up( Game* game, const uint_fast16_t w, const uint_fast16_t h);
+void pop_up( Game* game, const int w, const int h);
 void update_end_game( Game* game);
 
 void lineWinner(Game* game, const Coor dpt, const Coor arv);
@@ -20,6 +20,6 @@ void update_in_menu( Game* game);
 void update_in_menu_level( Game* game);
 
 void clickButtonMenu(Game* game, const bool Solo);
-void clickLevel(Game* game, const uint_fast16_t clickX, const uint_fast16_t clickY);
+void clickLevel(Game* game, const int clickX, const int clickY);
 
 #endif
